+++ sdk/src/main/java/ly/count/android/sdk/ConfigExperimental.java
@@ -1,0 +1,16 @@
+package ly.count.android.sdk;
+
+public class ConfigExperimental {
+    protected boolean viewNameRecordingEnabled = false;
+    protected boolean visibilityTrackingEnabled = false;
+
+    public ConfigExperimental enableViewNameRecording() {
+        viewNameRecordingEnabled = true;
+        return this;
+    }
+
+    public ConfigExperimental enableVisibilityTracking() {
+        visibilityTrackingEnabled = true;
+        return this;
+    }
+}