+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## XX.XX.XX
 * The Android SDK now supports Android 15 (API level 35)
+* The views will be stopped and restarted now while going to the background or foreground instead of resuming and pausing.
 
 ## 24.7.2
 * Mitigated an issue in the upload plugin that prevented the upload of a symbol file