+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## XX.XX.XX
 * The views will be stopped and restarted now while going to the background or foreground instead of resuming and pausing.
+* Added visibility tracking to the views and normal events.
 
 ## 24.7.2
 * Mitigated an issue in the upload plugin that prevented the upload of a symbol file