+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## XX.XX.XX
+* The views will be stopped and restarted now while going to the background or foreground instead of resuming and pausing.
+* The previous view name and the current view name have been added to the events. Additionally, the previous view name has been incorporated into the views. (Experimental!)
+* Added visibility tracking option to views and events. (Experimental!)
+
 ## 24.7.2
 * Mitigated an issue in the upload plugin that prevented the upload of a symbol file